+++ SpaceShip/src/SpaceShip.Domain/Entities/BaseEntity.cs
@@ -1,0 +1,17 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace SpaceShip.Domain.Entities;
+
+/// <summary>
+/// Base entity.
+/// </summary>
+public abstract class BaseEntity
+{
+    /// <summary>
+    /// Entity ID.
+    /// </summary>
+    [Key]
+    [Column("Id")]
+    public Guid Id { get; set; }
+}
