+++ SpaceShip/src/SpaceShip.Domain/Entities/Ship.cs
@@ -1,0 +1,34 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+using SpaceShip.Domain.Enums;
+
+namespace SpaceShip.Domain.Entities;
+
+/// <summary>
+/// Космический корабль.
+/// </summary>
+[Table("Ship")]
+public class Ship : BaseEntity
+{
+    /// <summary>
+    /// Имя корабля.
+    /// </summary>
+    [Column("Name")]
+    public string? Name { get; set; }
+
+    /// <summary>
+    /// Статус корабля.
+    /// </summary>
+    [Column("State")]
+    public SpaceshipState State { get; set; }
+
+    /// <summary>
+    /// Ход.
+    /// </summary>
+    [Column("Step")]
+    public short? Step { get; set; }
+
+    /// <summary>
+    /// Коллекция ресурсов.
+    /// </summary>
+    public virtual ICollection<Resource> Resources { get; set; }
+}
