+++ SpaceShip/src/SpaceShip.Domain/Enums/ResourceType.cs
@@ -1,0 +1,12 @@
+﻿namespace SpaceShip.Domain.Enums;
+
+/// <summary>
+/// Типы ресурсов.
+/// </summary>
+public enum ResourceType : byte
+{
+    Hull = 0,
+    ScrapMetal = 1,
+    Engine = 2,
+    Fuel = 3
+}
