+++ SpaceShip/src/SpaceShip.Domain/Enums/SpaceshipState.cs
@@ -1,0 +1,11 @@
+﻿namespace SpaceShip.Domain.Enums;
+
+/// <summary>
+/// State for spaceship models.
+/// </summary>
+public enum SpaceshipState : byte
+{
+    OK = 0,
+    Sleep = 1,
+    Finish = 2
+}
