+++ SpaceShip/src/SpaceShip.Domain/Implementation/BaseRepository.cs
@@ -1,122 +1,121 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SpaceShip.Domain.EfCore;
+using SpaceShip.Domain.Entities;
 using SpaceShip.Domain.Interfaces;
-using SpaceShip.Domain.Model;
 
-namespace SpaceShip.Domain.Implementation
+namespace SpaceShip.Domain.Implementation;
+
+/// <summary>
+/// Base repository.
+/// </summary>
+/// <typeparam name="T">Repository entity.</typeparam>
+public abstract class BaseRepository<T> : IRepository<T>
+    where T : BaseEntity, new()
 {
+    #region constructor
+
     /// <summary>
-    /// Base repository.
+    /// Constructor.
     /// </summary>
-    /// <typeparam name="T">Repository entity.</typeparam>
-    public abstract class BaseRepository<T> : IRepository<T>
-        where T : BaseEntity, new()
+    protected BaseRepository(EfCoreContext efCoreContext)
     {
-        #region constructor
+        Context = efCoreContext;
+        EntitySet = Context.Set<T>();
+    }
 
-        /// <summary>
-        /// Constructor.
-        /// </summary>
-        protected BaseRepository(EfCoreContext efCoreContext)
-        {
-            Context = efCoreContext;
-            EntitySet = Context.Set<T>();
-        }
+    #endregion
 
-        #endregion
+    #region protected fields
 
-        #region protected fields
+    /// <summary>
+    /// Context.
+    /// </summary>
+    protected EfCoreContext Context { get; }
 
-        /// <summary>
-        /// Context.
-        /// </summary>
-        protected EfCoreContext Context { get; }
+    /// <summary>
+    /// EntitySet.
+    /// </summary>
+    protected DbSet<T> EntitySet { get; }
 
-        /// <summary>
-        /// EntitySet.
-        /// </summary>
-        protected DbSet<T> EntitySet { get; }
+    #endregion
 
-        #endregion
+    #region public methods
 
-        #region public methods
+    /// <inheritdoc/>
+    public virtual T Create(bool saveChanges = false)
+    {
+        var entity = new T();
 
-        /// <inheritdoc/>
-        public virtual T Create(bool saveChanges = false)
-        {
-            var entity = new T();
+        var entityEntry = EntitySet.Add(entity);
 
-            var entityEntry = EntitySet.Add(entity);
-
-            if (saveChanges)
-            {
-                Context.SaveChanges();
-            }
-
-            return entityEntry.Entity;
-        }
-
-        /// <inheritdoc/>
-        public virtual T Create(T entity, bool saveChanges = false)
-        {
-            var entityEntry = EntitySet.Add(entity);
-
-            if (saveChanges)
-            {
-                Context.SaveChanges();
-            }
-
-            return entityEntry.Entity;
-        }
-
-        /// <inheritdoc/>
-        public virtual List<T> GetAll()
-        {
-            return EntitySet.ToList();
-        }
-
-        /// <inheritdoc/>
-        public virtual T? Get(Guid id)
-        {
-            return EntitySet.Find(id);
-        }
-
-        /// <inheritdoc/>
-        public void Update(T entity, bool saveChanges = false)
-        {
-            Context.Entry(entity).State = EntityState.Modified;
-
-            if (saveChanges)
-            {
-                Context.SaveChanges();
-            }
-        }
-
-        /// <inheritdoc/>
-        public bool Delete(Guid id, bool saveChanges = false)
-        {
-            var obj = EntitySet.Find(id);
-            if (obj == null)
-            {
-                return false;
-            }
-
-            EntitySet.Remove(obj);
-
-            if (saveChanges)
-            {
-                Context.SaveChanges();
-            }
-
-            return true;
-        }
-
-        /// <inheritdoc/>
-        public void SaveChanges()
+        if (saveChanges)
         {
             Context.SaveChanges();
         }
 
-        #endregion
+        return entityEntry.Entity;
     }
+
+    /// <inheritdoc/>
+    public virtual T Create(T entity, bool saveChanges = false)
+    {
+        var entityEntry = EntitySet.Add(entity);
+
+        if (saveChanges)
+        {
+            Context.SaveChanges();
+        }
+
+        return entityEntry.Entity;
+    }
+
+    /// <inheritdoc/>
+    public virtual List<T> GetAll()
+    {
+        return EntitySet.ToList();
+    }
+
+    /// <inheritdoc/>
+    public virtual T? Get(Guid id)
+    {
+        return EntitySet.Find(id);
+    }
+
+    /// <inheritdoc/>
+    public void Update(T entity, bool saveChanges = false)
+    {
+        Context.Entry(entity).State = EntityState.Modified;
+
+        if (saveChanges)
+        {
+            Context.SaveChanges();
+        }
+    }
+
+    /// <inheritdoc/>
+    public bool Delete(Guid id, bool saveChanges = false)
+    {
+        var obj = EntitySet.Find(id);
+        if (obj == null)
+        {
+            return false;
+        }
+
+        EntitySet.Remove(obj);
+
+        if (saveChanges)
+        {
+            Context.SaveChanges();
+        }
+
+        return true;
+    }
+
+    /// <inheritdoc/>
+    public void SaveChanges()
+    {
+        Context.SaveChanges();
+    }
+
+    #endregion
 }
