+++ SpaceShip/src/SpaceShip.Domain/Implementation/ResourceRepository.cs
@@ -1,24 +1,23 @@
 ﻿using SpaceShip.Domain.EfCore;
+using SpaceShip.Domain.Entities;
 using SpaceShip.Domain.Interfaces;
-using SpaceShip.Domain.Model;
 
-namespace SpaceShip.Domain.Implementation
+namespace SpaceShip.Domain.Implementation;
+
+/// <summary>
+/// Resource repository.
+/// </summary>
+public class ResourceRepository : BaseRepository<Resource>, IResourceRepository
 {
+    #region constructor
+
     /// <summary>
-    /// Resource repository.
+    /// Constructor.
     /// </summary>
-    public class ResourceRepository : BaseRepository<Resource>, IResourceRepository
+    public ResourceRepository(EfCoreContext context)
+        : base(context)
     {
-        #region constructor
-
-        /// <summary>
-        /// Constructor.
-        /// </summary>
-        public ResourceRepository(EfCoreContext context)
-            : base(context)
-        {
-        }
-
-        #endregion
     }
+
+    #endregion
 }
