--- SpaceShip/src/SpaceShip.Domain/Model/ResourceType.cs
@@ -1,13 +1,0 @@
-﻿namespace SpaceShip.Domain.Model
-{
-    /// <summary>
-    /// Типы ресурсов.
-    /// </summary>
-    public enum ResourceType : byte
-    {
-        Hull = 0,
-        ScrapMetal = 1,
-        Engine = 2,
-        Fuel = 3
-    }
-}
