--- SpaceShip/src/SpaceShip.Domain/Model/State/SpaceshipState.cs
@@ -1,12 +1,0 @@
-﻿namespace SpaceShip.Domain.Model.State
-{
-    /// <summary>
-    /// State for paceship models.
-    /// </summary>
-    public enum SpaceshipState : byte
-    {
-        OK = 0,
-        Sleep = 1,
-        Finish = 2
-    }
-}
